flaot 1024

+++ VITIS_HLS_SRC/specs.h
@@ -5,8 +5,10 @@
 #import <cstdint>
 #include "ap_int.h"
 
+//typedef uint8_t T;
+//typedef uint32_t Tout;
 typedef uint8_t T;
-typedef uint32_t Tout;
+typedef float Tout;
 typedef ap_uint<26> uint26_t;  // Tipo variable intermedia
 typedef ap_uint<23> uint23_t;  // Tipo variable intermedia
